+++ ws/client/connection.py
@@ -12,7 +12,6 @@ and making requests.
 
 import requests
 from requests.packages.urllib3.util.retry import Retry
-import ssl
 import http.cookiejar as cookielib
 import logging
 import copy
@@ -140,12 +139,9 @@ class Connection:
         session.auth = _auth
         session.params.update({"format": "json"})
 
-        # disallow TLS1.0 and TLS1.1, allow only TLS1.2 (and newer if suported
-        # by the used openssl version)
-        ssl_options = ssl.OP_NO_TLSv1 | ssl.OP_NO_TLSv1_1
         # granular control over requests' retries: https://stackoverflow.com/a/35504626
         retries = Retry(total=max_retries, backoff_factor=1, status_forcelist=[429, 500, 502, 503, 504])
-        adapter = TLSAdapter(ssl_options=ssl_options, max_retries=retries)
+        adapter = TLSAdapter(max_retries=retries)
         session.mount("https://", adapter)
         session.mount("http://", adapter)
         return session